+++ lib/model/Assistant.dart
@@ -239,7 +239,7 @@ enum MessageFeedback { good, bad }
 ///
 /// AssistantProvider
 ///
-enum AssistantProvider { uiuc, google, azure, grok }
+enum AssistantProvider { uiuc, google, grok }
 
 String? assistantProviderToKeyString(AssistantProvider? provider) {
   switch (provider) {
@@ -247,8 +247,6 @@ String? assistantProviderToKeyString(AssistantProvider? provider) {
       return 'uiuc';
     case AssistantProvider.google:
       return 'google';
-    case AssistantProvider.azure:
-      return 'azure';
     case AssistantProvider.grok:
       return 'grok';
     default:
@@ -262,8 +260,6 @@ String assistantProviderToDisplayString(AssistantProvider? provider) {
       return Localization().getStringEx('model.assistant.provider.uiuc.label', 'Illinois');
     case AssistantProvider.google:
       return Localization().getStringEx('model.assistant.provider.google.label', 'Google');
-    case AssistantProvider.azure:
-      return Localization().getStringEx('model.assistant.provider.azure.label', 'Azure');
     case AssistantProvider.grok:
       return Localization().getStringEx('model.assistant.provider.grok.label', 'Grok');
     default: